change admin ui

+++ app/Http/Controllers/Admin/AuthorController.php
@@ -4,66 +4,49 @@ namespace App\Http\Controllers\Admin;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-
-use App\Author;
-
+use Redirect,Response;
+use App\Models\Author;
+use App\Models\Book;
 
 class AuthorController extends Controller
 {
-    public function getIndex(){
-    	$allAuthor = Author::all();
-    	return view('admin\book\book_author_list', compact('allAuthor'));
+    public function index(){
+        $viewName = "tác giả";
+    	if(request()->ajax()) {
+            $authors = Author::select('*');
+
+            return datatables()->of($authors)
+            ->addColumn('action', function($authors){
+                $id = $authors->id;
+                return (string)view('admin/bookinfo/action', compact('id'));
+            })
+            ->rawColumns(['action'])
+            ->addIndexColumn()
+            ->make(true);
+        }
+    	return view('admin\bookinfo\bookinfo', compact('viewName'));
+    }
+      
+    public function edit($id)
+    {   
+        $where = array('id' => $id);
+        $author  = Author::where($where)->first();
+        return Response::json($author);
     }
 
-    public function postIndex(Request $request){
-        $this->validate($request, [
-            'value'=>'required|unique:author,name'
-        ], [
-            'value.required'=>'Bạn chưa nhập tên tác giả',
-            'value.unique'=>'Tác giả đã tồn tại'
-        ]);
-
-    	$author = new Author;
-    	$author->name = $request->value;
-    	$author->save();
-    	return redirect('admin/books/author')->with('message', 'Thêm tác giả thành công');
-    }
-
-    public function getEdit($id){
-    	$allAuthor = Author::all();
-    	$edit_author = Author::where('id', $id)->first();
-    	return view('admin\book\book_author_edit', compact('edit_author', 'allAuthor'));
-    }
-
-    public function postEdit(Request $request, $id){
-        $this->validate($request, [
-            'value'=>'required|unique:author,name'
-        ], [
-            'value.required'=>'Bạn chưa nhập tên tác giả',
-            'value.unique'=>'Tác giả đã tồn tại'
-        ]);
-
-    	$author = Author::find($id);
-    	$author->name = $request->value;
-    	$author->save();
-    	return redirect('admin/books/author');
+    public function store(Request $request)
+    {  
+        $data = Author::updateOrCreate(['id'=>$request->id], ['name'=>$request->name]);   
+        return Response::json($data);
     }
 
     public function delete($id){
-        $author = Author::find($id);
-        $author->delete();
+        $author = Author::where('id', '=', $id)->first();
 
-        return redirect('admin/books/author');
-    }
-
-    public function search(Request $request){
-        if($request->searchkey != ''){
-            $allAuthor = Author::where('name', 'like', "%$request->searchkey%")->get();
-            return view('admin\book\book_author_list', compact('allAuthor'));
+        if(count($author->Book) > 0){
+            return Response::json("error");
         }
-        return redirect('admin/books/author');
-    }
-    public function getSearch(){
-        return redirect('admin/books/author');
+        $author->delete();
+        return Response::json("success");  
     }
 }
